requirements.txt
pyscreenshot
Pillow
pytesseract
pyautogui